Add files via upload
A circular barplot is a barplot, with each bar displayed along a circle instead of a line. A circular bar chart is very ‘eye catching’ and allows better use of the space than a long usual barplot
In this figure, four different categories of plants are shown along with values and they are arranged from lowest to highest according to their values. In order to understand the values of the Y-axis, several lines are drawn to indicate the values in the Y-axis. The same information could also be plot using just basic bar plots. Circular bar plots make sense only if you have a huge number of bars to display, and if an obvious pattern pops out.