Add files via upload
A population pyramid is a graphical illustration of the distribution of a population (typically that of a country or region of the world) by age groups and gender; it typically forms the shape of a pyramid when the population is growing. Males are usually shown on the left and females on the right, and they may be measured by count or as a percentage of the total population.

The Figure is drawn with “R”. In this figure it shows the result of email marketing campaign. With the progress of each stage, lesser number of users were retained. Very minute work was done to prepare the data as the data contained negative values for males and positive values for females and the code handled the data accordingly. You can have a look at the code and try to figure this out what was done and let me know!