Add files via upload
A dendrogram is a network structure. It is constituted of a root node that gives birth to several nodes connected by edges or branches. The last nodes of the hierarchy are called leaves. It can also be linear or circular. The advantage of the circular version is that it uses the graphic space more efficiently:
The following figure was drawn using “R” with public source code. This figure shows the network between groups and subgroups and how they are linked to specific groups. The color indicates the different groups along with their clustering in different groups. The size of the colored ball indicates the values with a higher value having a bigger size.