Add files via upload
Diverging Bars is a barchart that can handle compare variation in values between small number of items (or categories) with respect to a fixed reference.
//
The Figure below is drawn using “R” with public source codes. The data used here is very famous “mtcars” dataset which is inbuilt in R. One of the column was given as miles per gallon for different car brands. These data were first normalized and then the diverging bars were plotted as below average of 0 and above average of zero.