Add files via upload
As someone pointed out to draw the map of Bangladesh using R. Here I have drawn the district level and divisional level image of Bangladesh. The code is fairly straightforward. The co-ordinates are publicly available on the internet and I have linked it in the code to download the data. There are some specific packages necessary and draw maps that can also be accessed through R. Hope you enjoy it.