Add files via upload
An animated bubble chart is the same as the bubble chart, but, you have to show how the values change over a fifth dimension (typically time). 
The figure is drawn using “R” with public source codes. The data is collected from gapminder package which contains the population, GDP, life expectancy across different years of all the countries of the world. Here I have separated countries based on color and grouped them by Africa, America, Asia, Europe, and Oceania which was available in the dataset already. For the animation purpose, there was a new package used called “gganimate”. Looks pretty cool !