Add files via upload
A 2d density plot is useful to study the relationship between 2 numeric variables if you have a huge number of points. To avoid overlapping (as in the scatterplot beside), it divides the plot area in a multitude of small fragment and represents the number of points in this fragment. One way to plot this data to divide the plot area in multiple hexagon and then plot it accordingly, so you get hexbin chart with layers of count indicated as color.
//
The figure is generated with random data with two particular variable, namely x and y. The data is also random. What it shows is the plotting of each point against each of the other variable. The hexagon ensures the overlap data and that is shown with intensity of the color. So the more data points are in the hexagonal region, the brighter the color is