Add files via upload
If you want to show the relationship as well as the distribution in the same chart, use the marginal histogram. It has a histogram of the X and Y variables at the margins of the scatterplot.
The following figure is drawn with “R” using public source code. The dataset is inbuilt mpg dataset. Here the city miles per gallon was plotted against highway miles per gallon for each car. The distribution of the data is shown as histogram in the margin of the figure. Give it a go!