Add files via upload
A donut or doughnut chart is a ring divided into sectors that each represent a proportion of the whole. It is very close to a pie chart.
//
The following figure is drawn with R. The data is randomly generated. The data contains three categorical variables A,B and C and counts of them. There are some modifications done of the data like counting percentage and cumulative values and also the position of the label for the figure. 
Code to Draw Figure