Add files via upload
Violin plots allow to visualization of the distribution of a numeric variable for one or several groups. They are very well adapted for large dataset
The figure is drawn using R. The dataset is collected from the github website which is a survey result. Performed some work on data preparation and then plotted the violin plot with ordering and on the horizontal axis.