Add files via upload
#One_Figure_A_Day_Challenge
#Day16
A slope chart (also called slope graph) shows developments between two points, e.g. between two dates. Slope charts can be seen as simplified line charts that just show the start and end point of each line. A slope graph can be used to show a 'before and after' story of different values, based on comparing their values at different points in time. The related values are connected by slopes.

The figure is drawn using R. The dataset is meanGDPperCapita of 5 continents in the year 1952 and 1957. The figure shows the change in meanGDPperCapita in them with green indicating growth and red indicating decline.