Add files via upload
As someone pointed me to do a Volcano plot, here it is. A volcano plot typically plots some measure of effect on the x-axis (typically the fold change) and the statistical significance on the y-axis (typically the -log10 of the p-value). Genes that are highly dysregulated are farther to the left and right sides, while highly significant changes appear higher on the plot.
The following figure is drawn with R. The main data for this figure is collected from an online platform that stores data from an experiment. The data that I have used here contains 3 columns which are gene name, log2foldchange and pvalue. The red and blue colored dots are upregulated and downregulated genes and their names are also shown in the figure. The code is very detailed as I was asked and will help anyone who wants to use it for their analysis.