Add files via upload
A heatmap is a graphical representation of data where the individual values contained in a matrix are represented as colors
The following figure is drawn with R. The data is an inbuilt time series dataset called presidents. This data contains the years, quarters, and rating of the presidents in those years. The heatmap shows a very nice summary of the data where in x axis you can see the years and y axis the quarters. The intensity of the color indicates the rating of t