Add files via upload
Waffle charts is a nice way of showing the categorical composition of the total population
The following figure is drawn with R. Though there is no direct function, it can be articulated by smartly maneuvering the ggplot2 using geom_tile() function. The data is an inbuilt mpg dataset.  The is rounded to accommodate in 10 by 10 waffle.