Add files via upload
A boxplot summarizes the distribution of a continuous variable. it is often criticized for hiding the underlying distribution of each group. Thus, showing individual observation using jitter on top of boxes is a good practice
The following figure is drawn with R. The data is randomly generated from some normally distributed data for each the category.